rename: Text.tsx -> index.tsx 변경 및 절대경로 수정
