+++ src/components/forms/Color/index.tsx
@@ -1,9 +1,8 @@
 import { useCallback } from "react";
 
-import { Text } from "@/components/typography/Text/Text";
-
-import { SelectedColor } from "@/components/forms/ColorPalette";
 import * as Styles from "@/components/forms/Color/index.style";
+import { SelectedColor } from "@/components/forms/ColorPalette";
+import { Text } from "@/components/typography/Text";
 
 export interface ColorProps extends SelectedColor {
     active?: boolean;
